Cut out the function of utils into another script
